# openapi.yaml
openapi: 3.0.0
info:
  title: AMP Email Generation API Platform
  version: 1.0.0
  description: |
    AI-powered AMP email template generation API with personalization, analytics, and integration support.
    
    ## Features
    - Generate AMP-compliant email templates from product URLs or data
    - Bulk batch processing with async job support
    - Personalization engine with merge tags
    - Real-time analytics and cost tracking
    - ESP integration (SendGrid, Resend, AWS SES)
    - Third-party AI agent compatibility
    
    ## Authentication
    All endpoints require API Key authentication via Bearer token in the Authorization header.
    
  contact:
    name: AMP Email Platform API Support
    url: https://amp-platform.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.amp-platform.com
    description: Production server
  - url: https://staging-api.amp-platform.com
    description: Staging server
  - url: http://localhost:3000
    description: Local development server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/analytics/campaign/{campaignId}:
    get:
      summary: Get campaign analytics
      description: Retrieve analytics and cost metrics for a specific campaign
      tags:
        - Analytics
      parameters:
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Campaign analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaign_id:
                    type: string
                  total_cost:
                    type: number
                  products_processed:
                    type: integer
                  templates_created:
                    type: integer
                  generation_time_ms:
                    type: integer
  /api/v1/analytics/company/{companyId}:
    get:
      summary: Get company analytics
      description: Retrieve aggregated analytics for a company
      tags:
        - Analytics
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Company analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  company_id:
                    type: string
                  total_campaigns:
                    type: integer
                  total_cost:
                    type: number
                  avg_cost_per_campaign:
                    type: number
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API_KEY
      description: API key obtained from platform dashboard
...